[development/file_reorganisation] moving filesystem files back into platform specific folders
